refactor: migrate Jest setup to jest.ts and remove deprecated mock files
