+++ packages/mvc/src/state.test.ts
@@ -1,8 +1,8 @@
+import { mockError, mockPromise } from '../jest';
 import { Context } from './context';
 import { get } from './instruction/get';
 import { ref } from './instruction/ref';
 import { set } from './instruction/set';
-import { mockError, mockPromise } from './mocks';
 import { State } from './state';
 
 it('will extend custom class', () => {
